requirements.txt
google-adk<1.16
fastapi<0.119
google-genai==1.49.0
